backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
Django==5.1.6
django-ratelimit==4.1.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
firebase-admin==6.7.0
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
img2pdf==0.6.0
inflection==0.5.1
joblib==1.4.2
keras-nightly==3.10.0.dev2025042703
libclang==18.1.1
lxml==5.3.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
multidict==6.4.3
namex==0.0.9
nltk==3.9.1
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pikepdf==9.7.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==14.0.0
rsa==4.9
setuptools==79.0.1
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
tb-nightly==2.19.0a20250218
tensorboard-data-server==0.7.2
termcolor==3.0.1
tf_nightly==2.20.0.dev20250425
tqdm==4.67.1
twilio==9.5.2
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
